[FlexibleHeader] Move all shift behavior APIs to their own header.
This is being done in an effort to isolate the shift behavior logic. This is being done in an effort to better document the separations of concerns of the FlexibleHeader's individual features.
